pyproject.toml
[tool.poetry]
name = "markov"
version = "0.1.0"
description = ""
authors = ["Alliana <56180684+Allianaab2m@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
markovify = "^0.9.4"
mecab-python = "^1.0.0"
unidic-lite = "^1.0.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
